moved iptables comment code to policy_linux.go
